18: Split RestConnection class out of HubIntRestService
