+++ src/main/java/com/blackducksoftware/integration/hub/HubIntRestService.java
@@ -22,26 +22,18 @@
 package com.blackducksoftware.integration.hub;
 
 import java.io.IOException;
-import java.lang.reflect.Constructor;
-import java.net.Authenticator;
-import java.net.PasswordAuthentication;
-import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.AbstractMap;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map.Entry;
 import java.util.Set;
 import java.util.regex.Pattern;
 
 import org.apache.commons.lang3.StringUtils;
-import org.restlet.Context;
 import org.restlet.Response;
 import org.restlet.data.ChallengeRequest;
-import org.restlet.data.ChallengeResponse;
 import org.restlet.data.CharacterSet;
 import org.restlet.data.Cookie;
-import org.restlet.data.CookieSetting;
 import org.restlet.data.MediaType;
 import org.restlet.data.Method;
 import org.restlet.data.Reference;
@@ -49,10 +41,10 @@ import org.restlet.engine.header.Header;
 import org.restlet.engine.header.HeaderConstants;
 import org.restlet.representation.StringRepresentation;
 import org.restlet.resource.ClientResource;
-import org.restlet.resource.ResourceException;
 import org.restlet.util.Series;
 
 import com.blackducksoftware.integration.hub.api.VersionComparison;
+import com.blackducksoftware.integration.hub.connection.RestConnection;
 import com.blackducksoftware.integration.hub.exception.BDRestException;
 import com.blackducksoftware.integration.hub.exception.HubIntegrationException;
 import com.blackducksoftware.integration.hub.exception.MissingPolicyStatusException;
@@ -80,72 +72,25 @@ import com.google.gson.reflect.TypeToken;
 
 public class HubIntRestService {
 
-    private Series<Cookie> cookies;
-
-    private final String baseUrl;
-
-    private int timeout = 120000;
+    private RestConnection hubConnection;
 
     private IntLogger logger;
 
-    private String proxyUsername;
-
-    private String proxyPassword;
-
-    private ClientResource reUsableResource;
-
     public HubIntRestService(final String baseUrl) throws URISyntaxException {
-	this.baseUrl = baseUrl;
-
+	hubConnection = new RestConnection(baseUrl);
     }
 
     public void setTimeout(final int timeout) {
-	if (timeout == 0) {
-	    throw new IllegalArgumentException(
-		    "Can not set the timeout to zero.");
-	}
-	// the User sets the timeout in seconds, so we translate to ms
-	this.timeout = timeout * 1000;
+	hubConnection.setTimeout(timeout);
     }
 
     public void setLogger(final IntLogger logger) {
 	this.logger = logger;
+	hubConnection.setLogger(logger);
     }
 
     public String getBaseUrl() {
-	return baseUrl;
-    }
-
-    private void attemptResetProxyCache() {
-	try {
-	    Class<?> sunAuthCacheValue;
-	    Class<?> sunAuthCache;
-	    Class<?> sunAuthCacheImpl;
-	    try {
-		sunAuthCacheValue = Class
-			.forName("sun.net.www.protocol.http.AuthCacheValue");
-		sunAuthCache = Class
-			.forName("sun.net.www.protocol.http.AuthCache");
-		sunAuthCacheImpl = Class
-			.forName("sun.net.www.protocol.http.AuthCacheImpl");
-	    } catch (final Exception e) {
-		// Must not be using a JDK with sun classes so we abandon this
-		// reset since it is sun specific
-		return;
-	    }
-
-	    final java.lang.reflect.Method m = sunAuthCacheValue
-		    .getDeclaredMethod("setAuthCache", sunAuthCache);
-
-	    final Constructor<?> authCacheImplConstr = sunAuthCacheImpl
-		    .getConstructor();
-	    final Object authCachImp = authCacheImplConstr.newInstance();
-
-	    m.invoke(null, authCachImp);
-
-	} catch (final Exception e) {
-	    logger.error(e.getMessage());
-	}
+	return hubConnection.getBaseUrl();
     }
 
     /**
@@ -157,134 +102,22 @@ public class HubIntRestService {
 	    final List<Pattern> noProxyHosts, final String proxyUsername,
 	    final String proxyPassword) {
 
-	cleanUpOldProxySettings();
-
-	if (!StringUtils.isBlank(proxyHost) && proxyPort > 0) {
-	    if (logger != null) {
-		logger.debug("Using Proxy : " + proxyHost + ", at Port : "
-			+ proxyPort);
-	    }
-
-	    System.setProperty("https.proxyHost", proxyHost);
-	    System.setProperty("https.proxyPort", Integer.toString(proxyPort));
-	    System.setProperty("http.proxyHost", proxyHost);
-	    System.setProperty("http.proxyPort", Integer.toString(proxyPort));
-
-	    if (!StringUtils.isBlank(proxyUsername)
-		    && !StringUtils.isBlank(proxyPassword)) {
-		this.proxyUsername = proxyUsername;
-		this.proxyPassword = proxyPassword;
-
-		// Java ignores http.proxyUser. Here's the workaround.
-		Authenticator.setDefault(new Authenticator() {
-		    // Need this to support digest authentication
-		    @Override
-		    protected PasswordAuthentication getPasswordAuthentication() {
-			if (getRequestorType() == RequestorType.PROXY) {
-			    return new PasswordAuthentication(proxyUsername,
-				    proxyPassword.toCharArray());
-			}
-			return null;
-		    }
-		});
-	    }
-	}
-	if (noProxyHosts != null && !noProxyHosts.isEmpty()) {
-	    String noProxyHostsString = null;
-	    for (final Pattern pattern : noProxyHosts) {
-		if (noProxyHostsString == null) {
-		    noProxyHostsString = pattern.toString();
-		} else {
-		    noProxyHostsString = noProxyHostsString + "|"
-			    + pattern.toString();
-		}
-	    }
-	    if (!StringUtils.isBlank(noProxyHostsString)) {
-		System.setProperty("http.nonProxyHosts", noProxyHostsString);
-	    }
-	}
+	hubConnection.setProxyProperties(proxyHost, proxyPort, noProxyHosts,
+		proxyUsername, proxyPassword);
     }
 
     public ClientResource createClientResource() throws URISyntaxException {
-	return createClientResource(getBaseUrl());
+	return hubConnection.createClientResource(getBaseUrl());
     }
 
     public ClientResource createClientResource(final String providedUrl)
 	    throws URISyntaxException {
-
-	final Context context = new Context();
-
-	// the socketTimeout parameter is used in the httpClient extension that
-	// we do not use
-	// We can probably remove this parameter
-	final String stringTimeout = String.valueOf(timeout);
-
-	context.getParameters().add("socketTimeout", stringTimeout);
-
-	context.getParameters().add("socketConnectTimeoutMs", stringTimeout);
-	context.getParameters().add("readTimeout", stringTimeout);
-	// Should throw timeout exception after the specified timeout, default
-	// is 2 minutes
-
-	final ClientResource resource = new ClientResource(context, new URI(
-		providedUrl));
-	resource.getRequest().setCookies(getCookies());
-	return resource;
-    }
-
-    /**
-     * Clears the previously set System properties I.E. https.proxyHost,
-     * https.proxyPort, http.proxyHost, http.proxyPort, http.nonProxyHosts
-     *
-     */
-    private void cleanUpOldProxySettings() {
-	System.clearProperty("https.proxyHost");
-	System.clearProperty("https.proxyPort");
-	System.clearProperty("http.proxyHost");
-	System.clearProperty("http.proxyPort");
-	System.clearProperty("http.nonProxyHosts");
-
-	attemptResetProxyCache();
-
-	Authenticator.setDefault(null);
+	return hubConnection.createClientResource(providedUrl);
     }
 
     public void parseChallengeRequestRawValue(
 	    final ChallengeRequest proxyChallengeRequest) {
-	if (proxyChallengeRequest == null
-		|| StringUtils.isBlank(proxyChallengeRequest.getRawValue())) {
-	    return;
-	}
-	final String rawValue = proxyChallengeRequest.getRawValue();
-
-	final String[] splitRawValue = rawValue.split(",");
-	for (final String currentValue : splitRawValue) {
-	    final String trimmedCurrentValue = currentValue.trim();
-	    if (StringUtils.isBlank(proxyChallengeRequest.getRealm())
-		    && trimmedCurrentValue.startsWith("realm=")) {
-		final String realm = trimmedCurrentValue.substring("realm="
-			.length());
-		proxyChallengeRequest.setRealm(realm);
-	    } else if (StringUtils.isBlank(proxyChallengeRequest
-		    .getServerNonce())
-		    && trimmedCurrentValue.startsWith("nonce=")) {
-		final String nonce = trimmedCurrentValue.substring("nonce="
-			.length());
-		proxyChallengeRequest.setServerNonce(nonce);
-	    } else if ((proxyChallengeRequest.getQualityOptions() == null || proxyChallengeRequest
-		    .getQualityOptions().isEmpty())
-		    && trimmedCurrentValue.startsWith("qop=")) {
-		final String qop = trimmedCurrentValue.substring("qop="
-			.length());
-		final List<String> qualityOptions = new ArrayList<String>();
-		qualityOptions.add(qop);
-		proxyChallengeRequest.setQualityOptions(qualityOptions);
-	    } else if (trimmedCurrentValue.startsWith("stale=")) {
-		final String stale = trimmedCurrentValue.substring("stale="
-			.length());
-		proxyChallengeRequest.setStale(Boolean.valueOf(stale));
-	    }
-	}
+	hubConnection.parseChallengeRequestRawValue(proxyChallengeRequest);
     }
 
     private void logMessage(final LogLevel level, final String txt) {
@@ -311,176 +144,11 @@ public class HubIntRestService {
     public int setCookies(final String hubUserName, final String hubPassword)
 	    throws HubIntegrationException, URISyntaxException, BDRestException {
 
-	final ClientResource resource = createClientResource();
-	resource.addSegment("j_spring_security_check");
-	resource.setMethod(Method.POST);
-
-	final StringRepresentation stringRep = new StringRepresentation(
-		"j_username=" + hubUserName + "&j_password=" + hubPassword);
-	stringRep.setCharacterSet(CharacterSet.UTF_8);
-	stringRep.setMediaType(MediaType.APPLICATION_WWW_FORM);
-	resource.getRequest().setEntity(stringRep);
-
-	logMessage(LogLevel.TRACE, "Cookies before auth : ");
-	if (cookies != null) {
-	    for (final Cookie ck : cookies) {
-		logMessage(
-			LogLevel.TRACE,
-			"Cookie, name = " + ck.getName() + " , domain = "
-				+ ck.getDomain() + " , path = " + ck.getPath()
-				+ " , value = " + ck.getValue()
-				+ " , version = " + ck.getVersion());
-	    }
-	} else {
-	    logMessage(LogLevel.TRACE,
-		    "Current 'Cookies' is null (none have ever been set yet).");
-	}
-
-	logMessage(LogLevel.TRACE, "Resource : " + resource.toString());
-	logMessage(LogLevel.TRACE, "Request : "
-		+ resource.getRequest().toString());
-
-	if (!resource.getRequest().getAttributes().isEmpty()) {
-	    logMessage(LogLevel.TRACE, "Request attributes : ");
-	    for (final Entry<String, Object> requestAtt : resource.getRequest()
-		    .getAttributes().entrySet()) {
-		logMessage(LogLevel.TRACE,
-			"Attribute key : " + requestAtt.getKey());
-		logMessage(LogLevel.TRACE,
-			"Attribute value : " + requestAtt.getValue());
-		logMessage(LogLevel.TRACE, "");
-	    }
-	    logMessage(LogLevel.TRACE, "Request headers : ");
-	    final Series<Header> requestheaders = (Series<Header>) resource
-		    .getRequest().getAttributes()
-		    .get(HeaderConstants.ATTRIBUTE_HEADERS);
-	    if (requestheaders != null) {
-		logMessage(LogLevel.TRACE, "Request headers : ");
-		for (final Header header : requestheaders) {
-		    if (null == header) {
-			logMessage(LogLevel.TRACE, "received a null header");
-		    } else {
-			logMessage(LogLevel.TRACE,
-				"Header name : " + header.getName());
-			logMessage(LogLevel.TRACE,
-				"Header value : " + header.getValue());
-			logMessage(LogLevel.TRACE, "");
-		    }
-		}
-	    } else {
-		logMessage(LogLevel.TRACE, "Request headers : NONE");
-	    }
-	} else {
-	    logMessage(LogLevel.TRACE,
-		    "Request does not have any attributes/headers.");
-	}
-
-	handleRequest(resource, null, 0);
-
-	logMessage(LogLevel.TRACE, "Response : "
-		+ resource.getResponse().toString());
-
-	if (!resource.getResponse().getAttributes().isEmpty()) {
-	    logMessage(LogLevel.TRACE, "Response attributes : ");
-	    for (final Entry<String, Object> requestAtt : resource
-		    .getResponse().getAttributes().entrySet()) {
-		logMessage(LogLevel.TRACE,
-			"Attribute key : " + requestAtt.getKey());
-		logMessage(LogLevel.TRACE,
-			"Attribute value : " + requestAtt.getValue());
-		logMessage(LogLevel.TRACE, "");
-	    }
-	    final Series<Header> responseheaders = (Series<Header>) resource
-		    .getResponse().getAttributes()
-		    .get(HeaderConstants.ATTRIBUTE_HEADERS);
-	    if (responseheaders != null) {
-		logMessage(LogLevel.TRACE, "Response headers : ");
-		for (final Header header : responseheaders) {
-		    if (null == header) {
-			logMessage(LogLevel.TRACE, "received a null header");
-		    } else {
-			logMessage(LogLevel.TRACE,
-				"Header name : " + header.getName());
-			logMessage(LogLevel.TRACE,
-				"Header value : " + header.getValue());
-			logMessage(LogLevel.TRACE, "");
-		    }
-		}
-	    } else {
-		logMessage(LogLevel.TRACE, "Response headers : NONE");
-	    }
-	} else {
-	    logMessage(LogLevel.TRACE,
-		    "Response does not have any attributes/headers.");
-	}
-
-	logMessage(LogLevel.TRACE, "Status Code : "
-		+ resource.getResponse().getStatus().getCode());
-
-	final int statusCode = resource.getResponse().getStatus().getCode();
-	if (statusCode == 204) {
-	    final Series<CookieSetting> cookieSettings = resource.getResponse()
-		    .getCookieSettings();
-	    if (cookieSettings != null) {
-		logMessage(LogLevel.TRACE, "Set-Cookies returned : "
-			+ cookieSettings.size());
-	    } else {
-		logMessage(LogLevel.TRACE, "Set-Cookies returned : NULL");
-	    }
-
-	    final Series<Cookie> requestCookies = resource.getRequest()
-		    .getCookies();
-	    if (cookieSettings != null && !cookieSettings.isEmpty()) {
-		for (final CookieSetting ck : cookieSettings) {
-		    if (ck == null) {
-			continue;
-		    }
-		    logMessage(LogLevel.TRACE,
-			    "Set-Cookie, name = " + ck.getName()
-				    + " , domain = " + ck.getDomain()
-				    + " , path = " + ck.getPath()
-				    + " , value = " + ck.getValue()
-				    + " , version = " + ck.getVersion());
-
-		    final Cookie cookie = new Cookie();
-		    cookie.setName(ck.getName());
-		    cookie.setDomain(ck.getDomain());
-		    cookie.setPath(ck.getPath());
-		    cookie.setValue(ck.getValue());
-		    cookie.setVersion(ck.getVersion());
-		    requestCookies.add(cookie);
-		}
-	    }
-	    if (requestCookies == null || requestCookies.size() == 0) {
-		throw new HubIntegrationException(
-			"Could not establish connection to '" + getBaseUrl()
-				+ "' . Failed to retrieve cookies");
-	    }
-
-	    cookies = requestCookies;
-
-	    logMessage(LogLevel.TRACE, "Cookies after auth : ");
-	    if (cookies != null) {
-		for (final Cookie ck : cookies) {
-		    logMessage(LogLevel.TRACE, "Cookie, name = " + ck.getName()
-			    + " , domain = " + ck.getDomain() + " , path = "
-			    + ck.getPath() + " , value = " + ck.getValue()
-			    + " , version = " + ck.getVersion());
-		}
-	    } else {
-		logMessage(LogLevel.TRACE, "New 'Cookies' are null.");
-	    }
-	} else {
-	    throw new HubIntegrationException(resource.getResponse()
-		    .getStatus().toString());
-	}
-	reUsableResource = createClientResource(); // Now that cookies are set,
-						   // we can create this
-	return resource.getResponse().getStatus().getCode();
+	return hubConnection.setCookies(hubUserName, hubPassword);
     }
 
     public Series<Cookie> getCookies() {
-	return cookies;
+	return hubConnection.getCookies();
     }
 
     /**
@@ -495,7 +163,7 @@ public class HubIntRestService {
 	resource.addQueryParameter("q", "name:" + projectName);
 	resource.addQueryParameter("limit", "15");
 	resource.setMethod(Method.GET);
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 	final int responseCode = resource.getResponse().getStatus().getCode();
 
 	if (RestletUtil.isSuccess(responseCode)) {
@@ -529,7 +197,7 @@ public class HubIntRestService {
 	resource.addQueryParameter("q", "name:" + projectName);
 	resource.addQueryParameter("limit", "15");
 	resource.setMethod(Method.GET);
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 	final int responseCode = resource.getResponse().getStatus().getCode();
 
 	if (RestletUtil.isSuccess(responseCode)) {
@@ -567,7 +235,7 @@ public class HubIntRestService {
 	    BDRestException, URISyntaxException {
 	final ClientResource resource = createClientResource(projectUrl);
 	resource.setMethod(Method.GET);
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 	final int responseCode = resource.getResponse().getStatus().getCode();
 
 	if (RestletUtil.isSuccess(responseCode)) {
@@ -588,7 +256,7 @@ public class HubIntRestService {
 	    throws IOException, BDRestException, URISyntaxException {
 	final ClientResource resource = createClientResource(versionUrl);
 	resource.setMethod(Method.GET);
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 	final int responseCode = resource.getResponse().getStatus().getCode();
 
 	if (RestletUtil.isSuccess(responseCode)) {
@@ -633,7 +301,7 @@ public class HubIntRestService {
 		.getLink(ProjectItem.VERSION_LINK));
 	resource.addQueryParameter("limit", "10000000");
 	resource.setMethod(Method.GET);
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 	final int responseCode = resource.getResponse().getStatus().getCode();
 
 	if (RestletUtil.isSuccess(responseCode)) {
@@ -675,7 +343,7 @@ public class HubIntRestService {
 	stringRep.setMediaType(MediaType.APPLICATION_JSON);
 	stringRep.setCharacterSet(CharacterSet.UTF_8);
 	resource.getRequest().setEntity(stringRep);
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 	final int responseCode = resource.getResponse().getStatus().getCode();
 
 	if (responseCode == 201) {
@@ -734,7 +402,7 @@ public class HubIntRestService {
 	stringRep.setMediaType(MediaType.APPLICATION_JSON);
 	stringRep.setCharacterSet(CharacterSet.UTF_8);
 	resource.getRequest().setEntity(stringRep);
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 	responseCode = resource.getResponse().getStatus().getCode();
 
 	if (responseCode == 201) {
@@ -780,7 +448,7 @@ public class HubIntRestService {
 	int responseCode = 0;
 
 	resource.setMethod(Method.GET);
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 	responseCode = resource.getResponse().getStatus().getCode();
 
 	if (RestletUtil.isSuccess(responseCode)) {
@@ -809,7 +477,7 @@ public class HubIntRestService {
 	int responseCode = 0;
 
 	resource.setMethod(Method.GET);
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 	responseCode = resource.getResponse().getStatus().getCode();
 
 	if (RestletUtil.isSuccess(responseCode)) {
@@ -861,7 +529,7 @@ public class HubIntRestService {
 
 	    resource.setMethod(Method.GET);
 
-	    handleRequest(resource, null, 0);
+	    hubConnection.handleRequest(resource, null, 0);
 
 	    final int responseCode = resource.getResponse().getStatus()
 		    .getCode();
@@ -943,7 +611,7 @@ public class HubIntRestService {
 	stringRep.setMediaType(MediaType.APPLICATION_JSON);
 	stringRep.setCharacterSet(CharacterSet.UTF_8);
 	resource.getRequest().setEntity(stringRep);
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 
 	final int responseCode = resource.getResponse().getStatus().getCode();
 
@@ -981,7 +649,7 @@ public class HubIntRestService {
 
 	final ClientResource resource = createClientResource(reportUrl);
 	resource.setMethod(Method.DELETE);
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 
 	final int responseCode = resource.getResponse().getStatus().getCode();
 	if (responseCode != 204) {
@@ -1013,7 +681,7 @@ public class HubIntRestService {
 
 	resource.setMethod(Method.GET);
 
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 	final int responseCode = resource.getResponse().getStatus().getCode();
 
 	if (responseCode == 200) {
@@ -1040,7 +708,7 @@ public class HubIntRestService {
 
 	resource.setMethod(Method.GET);
 
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 	final int responseCode = resource.getResponse().getStatus().getCode();
 
 	if (responseCode == 200) {
@@ -1081,7 +749,7 @@ public class HubIntRestService {
 
 	resource.setMethod(Method.GET);
 
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 
 	final int responseCode = resource.getResponse().getStatus().getCode();
 
@@ -1114,7 +782,7 @@ public class HubIntRestService {
 
 	resource.setMethod(Method.GET);
 
-	handleRequest(resource, null, 0);
+	hubConnection.handleRequest(resource, null, 0);
 	final int responseCode = resource.getResponse().getStatus().getCode();
 
 	if (responseCode == 200) {
@@ -1150,9 +818,10 @@ public class HubIntRestService {
 	    throws ResourceDoesNotExistException, URISyntaxException,
 	    IOException {
 
+	// TODO simplify
 	Reference queryRef = RestletUtil.createReference(url);
-	ClientResource resource = RestletUtil.getResource(reUsableResource,
-		queryRef);
+	ClientResource resource = RestletUtil.getResource(
+		hubConnection.createClientResource(), queryRef);
 
 	logMessage(LogLevel.DEBUG, "Resource: " + resource);
 	int responseCode = RestletUtil.getResponseStatusCode(resource);
@@ -1188,10 +857,11 @@ public class HubIntRestService {
 	    throws IOException, ResourceDoesNotExistException,
 	    URISyntaxException {
 
+	// TODO simplify
 	Reference queryRef = RestletUtil.createReference(getBaseUrl(),
 		urlSegments, queryParameters);
-	ClientResource resource = RestletUtil.getResource(reUsableResource,
-		queryRef);
+	ClientResource resource = RestletUtil.getResource(
+		hubConnection.createClientResource(), queryRef);
 
 	logMessage(LogLevel.DEBUG, "Resource: " + resource);
 	int responseCode = RestletUtil.getResponseStatusCode(resource);
@@ -1204,45 +874,6 @@ public class HubIntRestService {
 			    + urlSegments + " and query parameters "
 			    + queryParameters + "; errorCode: " + responseCode
 			    + "; " + resource.toString(), resource);
-	}
-    }
-
-    private void handleRequest(final ClientResource resource,
-	    final ChallengeRequest proxyChallengeRequest, final int attempt)
-	    throws BDRestException {
-	if (proxyChallengeRequest != null) {
-	    // This should replace the authenticator for the proxy
-	    // authentication
-	    // BUT it doesn't work for Digest authentication
-	    parseChallengeRequestRawValue(proxyChallengeRequest);
-	    resource.setProxyChallengeResponse(new ChallengeResponse(
-		    proxyChallengeRequest.getScheme(), null, proxyUsername,
-		    proxyPassword.toCharArray(), null, proxyChallengeRequest
-			    .getRealm(), null, null, proxyChallengeRequest
-			    .getDigestAlgorithm(), null, null,
-		    proxyChallengeRequest.getServerNonce(), 0, 0L));
-	}
-	try {
-	    resource.handle();
-	} catch (final ResourceException e) {
-	    if (resource.getProxyChallengeRequests() != null
-		    && !resource.getProxyChallengeRequests().isEmpty()
-		    && StringUtils.isNotBlank(proxyUsername)
-		    && StringUtils.isNotBlank(proxyPassword)) {
-
-		final ChallengeRequest newChallengeRequest = resource
-			.getProxyChallengeRequests().get(0);
-		if (attempt < 2) {
-		    handleRequest(resource, newChallengeRequest, attempt + 1);
-		} else {
-		    throw new BDRestException(
-			    "Too many proxy authentication attempts.", e,
-			    resource);
-		}
-	    }
-	    throw new BDRestException(
-		    "Problem connecting to the Hub server provided.", e,
-		    resource);
 	}
     }
